+++ e2e/runner/runner.go
@@ -70,14 +70,18 @@ type E2ERunner struct {
 	BTCDeployerAddress    *btcutil.AddressWitnessPubKeyHash
 	SolanaDeployerAddress solana.PublicKey
 
+	// all clients.
+	// a reference to this type is required to enable creating a new E2ERunner.
+	Clients Clients
+
 	// rpc clients
 	ZEVMClient   *ethclient.Client
 	EVMClient    *ethclient.Client
 	BtcRPCClient *rpcclient.Client
 	SolanaClient *rpc.Client
 
-	// grpc clients
-	AutorithyClient   authoritytypes.QueryClient
+	// zetacored grpc clients
+	AuthorityClient   authoritytypes.QueryClient
 	CctxClient        crosschaintypes.QueryClient
 	FungibleClient    fungibletypes.QueryClient
 	AuthClient        authtypes.QueryClient
@@ -165,19 +169,7 @@ func NewE2ERunner(
 	name string,
 	ctxCancel context.CancelFunc,
 	account config.Account,
-	evmClient *ethclient.Client,
-	zevmClient *ethclient.Client,
-	authorityClient authoritytypes.QueryClient,
-	cctxClient crosschaintypes.QueryClient,
-	fungibleClient fungibletypes.QueryClient,
-	authClient authtypes.QueryClient,
-	bankClient banktypes.QueryClient,
-	observerClient observertypes.QueryClient,
-	lightclientClient lightclienttypes.QueryClient,
-	evmAuth *bind.TransactOpts,
-	zevmAuth *bind.TransactOpts,
-	btcRPCClient *rpcclient.Client,
-	solanaClient *rpc.Client,
+	clients Clients,
 	logger *Logger,
 	opts ...E2ERunnerOption,
 ) *E2ERunner {
@@ -187,20 +179,22 @@ func NewE2ERunner(
 
 		Account: account,
 
-		ZEVMClient:        zevmClient,
-		EVMClient:         evmClient,
-		AutorithyClient:   authorityClient,
-		CctxClient:        cctxClient,
-		FungibleClient:    fungibleClient,
-		AuthClient:        authClient,
-		BankClient:        bankClient,
-		ObserverClient:    observerClient,
-		LightclientClient: lightclientClient,
+		Clients: clients,
 
-		EVMAuth:      evmAuth,
-		ZEVMAuth:     zevmAuth,
-		BtcRPCClient: btcRPCClient,
-		SolanaClient: solanaClient,
+		ZEVMClient:        clients.Zevm,
+		EVMClient:         clients.Evm,
+		AuthorityClient:   clients.Zetacore.Authority,
+		CctxClient:        clients.Zetacore.Crosschain,
+		FungibleClient:    clients.Zetacore.Fungible,
+		AuthClient:        clients.Zetacore.Auth,
+		BankClient:        clients.Zetacore.Bank,
+		ObserverClient:    clients.Zetacore.Observer,
+		LightclientClient: clients.Zetacore.Lightclient,
+
+		EVMAuth:      clients.EvmAuth,
+		ZEVMAuth:     clients.ZevmAuth,
+		BtcRPCClient: clients.BtcRPC,
+		SolanaClient: clients.Solana,
 
 		Logger: logger,
 	}