+++ changelog.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## v20.0.0
 
 ### Features